Vboot wrapper initial implementation
Patch 1: Initial change
Patch 2: Fix comment in vboot_struct.h
Patch 3: Revert files unintentionally reverted
Patch 4: (rebase)
Patch 5: (rebase)
Patch 6: Revert files unintentionally reverted (again)
Patch 7: Fix mocked tlcl for ARM build

BUG=chromium-os:17010
TEST=make && make runtests; works on H2C; emerge-tegra2_seaboard chromeos-bootimage compiles

Reviewed-on: http://gerrit.chromium.org/gerrit/3593
Tested-by: Randall Spangler <rspangler@chromium.org>

+++ firmware/include/load_firmware_fw.h
@@ -100,7 +100,7 @@ int LoadFirmware(LoadFirmwareParams* params);
  * by [size] bytes stored in [*data].  This function must only be
  * called inside GetFirmwareBody(). */
 void UpdateFirmwareBodyHash(LoadFirmwareParams* params,
-                            uint8_t* data, uint64_t size);
+                            uint8_t* data, uint32_t size);
 
 /* Handle S3 resume.
  *
